fix(driving-learners-permit): remove FF and minor cleanup
