+++ libs/application/templates/driving-learners-permit/src/graphql/index.ts
@@ -1,4 +1,5 @@
 import { gql } from '@apollo/client'
+
 export const IDENTITY_QUERY = gql`
   query IdentityQuery($input: IdentityInput!) {
     identity(input: $input) {
@@ -15,6 +16,18 @@ export const LOOKUP_STUDENT_QUERY = gql`
     drivingLicenseStudentCanGetPracticePermit(input: $input) {
       errorCode
       isOk
+    }
+  }
+`
+
+export const ELIGIBILITY_QUERY = gql`
+  query EligibilityQuery {
+    learnerMentorEligibility {
+      isEligible
+      requirements {
+        key
+        requirementMet
+      }
     }
   }
 `
