+++ libs/application/templates/driving-learners-permit/src/lib/constants.ts
@@ -1,6 +1,20 @@
 import { Application, DefaultEvents } from '@island.is/application/types'
 import { MessageDescriptor } from 'react-intl'
 
+export enum ApiActions {
+  createApplication = 'createApplication',
+  completeApplication = 'completeApplication',
+}
+
+export enum FakeDataFeature {
+  allowFake = 'applicationTemplateDrivingLearnersPermitAllowFakeData',
+}
+
+export const B_FULL = 'B-full'
+export const B_TEMP = 'B-temp'
+
+export type DrivingLicenseApplicationFor = typeof B_FULL | typeof B_TEMP
+
 export type Events =
   | { type: DefaultEvents.SUBMIT }
   | { type: DefaultEvents.PAYMENT }
