--- libs/application/templates/driving-learners-permit/src/lib/utils/formUtils.ts
@@ -1,6 +1,0 @@
-import { getValueViaPath } from '@island.is/application/core'
-import { FormValue } from '@island.is/application/types'
-import { YES } from '../constants'
-
-export const allowFakeCondition = (result = YES) => (answers: FormValue) =>
-  getValueViaPath(answers, 'fakeData.useFakeData') === result
