--- libs/application/templates/driving-learners-permit/src/shared/constants.ts
@@ -1,13 +1,0 @@
-export enum ApiActions {
-  createApplication = 'createApplication',
-  completeApplication = 'completeApplication',
-}
-
-export enum FakeDataFeature {
-  allowFake = 'applicationTemplateDrivingLearnersPermitAllowFakeData',
-}
-
-export const B_FULL = 'B-full'
-export const B_TEMP = 'B-temp'
-
-export type DrivingLicenseApplicationFor = typeof B_FULL | typeof B_TEMP
