--- libs/application/templates/driving-learners-permit/src/shared/index.ts
@@ -1,1 +1,0 @@
-export * from './constants'
