--- libs/application/templates/driving-learners-permit/src/shared/utils.ts
@@ -1,8 +1,0 @@
-import { FeatureFlagClient } from '@island.is/feature-flags'
-
-export async function truthyFeatureFromClient(
-  featureFlagClient: FeatureFlagClient,
-  key: string,
-): Promise<boolean> {
-  return !!(await featureFlagClient.getValue(key, false))
-}
